Refactor with updated dependencies
