+++ .github/workflows/unit-tests.yaml
@@ -22,17 +22,26 @@ jobs:
           key: ${{ runner.os }}-pip-${{ hashFiles('**/requirements.txt') }}
           restore-keys: |
             ${{ runner.os }}-pip-
-      - name: Set up Python '3.9'
+      - name: Set up Python '3.10'
         uses: actions/setup-python@v4
         with:
-          python-version: '3.9'
+          python-version: '3.10'
 
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
           pip install -r requirements-dev.txt
           pip install pytest
+
+      - name: Start PostgreSQL
+        run: |
+          sudo systemctl start postgresql.service
+          pg_isready
+          sudo -u postgres psql --command="CREATE USER testuser PASSWORD 'password'" --command="\du"
+          sudo -u postgres createdb idva_token_test --owner=testuser
       - name: Test with Pytest unit tests
         run: |
           export APP_SETTINGS="server.config.LocalDevConfig"
+          export DB_USER: testuser
+          export DB_PASSWORD: password
           python -m pytest
