+++ runtime.txt
@@ -1,1 +1,1 @@
-python-3.9.14
+python-3.10.1
