--- server/auth/__init__.py
@@ -1,1 +1,0 @@
-# server/auth/__init__.py
