--- tests/helpers.py
@@ -1,1 +1,0 @@
-# tests/helpers.py
